+++ services/cardiology-cloud/cardiology-session/src/main/java/com/sherwinzeng/cardiology/cardiologysession/request/CheckpointDeleteRequestParams.java
@@ -4,14 +4,11 @@ import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 
 @Data
-public class GeneralUnderstandingRequestParams {
+public class CheckpointDeleteRequestParams {
 
     @NotBlank(message = "uid 不能为空")
     private String uid;
 
     @NotBlank(message = "session 不能为空")
     private String session;
-
-    @NotBlank(message = "message 不能为空")
-    private String message;
 }